fixing tests issue with circleCi
