+++ tests/Unit/ExampleTest.php
@@ -1,0 +1,16 @@
+<?php
+namespace Tests\Unit;
+use Tests\TestCase;
+use Illuminate\Foundation\Testing\RefreshDatabase;
+class ExampleTest extends TestCase
+{
+    /**
+     * A basic test example.
+     *
+     * @return void
+     */
+    public function testBasicTest()
+    {
+        $this->assertTrue(true);
+    }
+}
